Task 16

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -21,21 +21,21 @@ int main(void)
 void fibonacci(void)
 {
 	const unsigned long max = 98;
-	unsigned long i, n1, n2, fib;
+	double i, n1, n2, fib;
 
 	n2 = 2;
 	n1 = 1;
-	printf("\n%lu, %lu, ", n1, n2);
+	printf("\n%.0f, %.0f, ", n1, n2);
 	for (i = 2; i <= max; i++)
 	{
 		fib = n2 + n1;
 		if (i == max)
 		{
-			printf("%lu", fib);
+			printf("%.0f", fib);
 		}
 		else
 		{
-			printf("%lu, ", fib);
+			printf("%.0f, ", fib);
 		}
 		n1 = n2;
 		n2 = fib;
